Change data type to Integer

+++ src/main/java/com/smallworld/data/Sender.java
@@ -5,5 +5,5 @@ import lombok.Data;
 @Data
 public class Sender {
     private String fullName;
-    private Byte age;
+    private Integer age;
 }
